MBY: rename the component folder
