+++ src/app/app.component.ts
@@ -1,16 +1,15 @@
+import { AboutComponent } from "./page-sections/about/about.component";
 import { Component } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
-import { HeaderComponent } from "./components/header/header.component";
-import { HeroComponent } from "./components/hero/hero.component";
-import { AboutComponent } from "./components/about/about.component";
-import { TechnologiesComponent } from "./components/technologies/technologies.component";
-import { ExperienceComponent } from "./components/experience/experience.component";
-import { ProjectsComponent } from "./components/projects/projects.component";
-import { ContactComponent } from "./components/contact/contact.component";
-import { TestimonialComponent } from "./components/testimonial/testimonial.component";
-import { ServicesComponent } from "./components/services/services.component";
+import { ContactComponent } from "./page-sections/contact/contact.component";
 import { ContactModalComponent } from "./shared/contact-modal/contact-modal.component";
-
+import { ExperienceComponent } from "./page-sections/experience/experience.component";
+import { HeaderComponent } from "./page-sections/header/header.component";
+import { HeroComponent } from "./page-sections/hero/hero.component";
+import { ProjectsComponent } from "./page-sections/projects/projects.component";
+import { RouterOutlet } from '@angular/router';
+import { ServicesComponent } from "./page-sections/services/services.component";
+import { TechnologiesComponent } from "./page-sections/technologies/technologies.component";
+import { TestimonialComponent } from "./page-sections/testimonial/testimonial.component";
 
 @Component({
   selector: 'app-root',
