--- src/app/components/about/about.component.html
@@ -1,17 +1,0 @@
-<div class="border-b border-neutral-900 pb-4">
-  <h1 class="my-20 text-center text-4xl">About
-    <span class="text-neutral-500"> Me</span>
-  </h1>
-  <div class="flex flex-wrap">
-    <div class="w-full lg:w-1/2 lg:p-8">
-      <div class="flex items-center justify-center">
-        <img class="rounded-2xl" src="/assets/images/about.jpg" alt="abotMeImage">
-      </div>
-    </div>
-    <div class="w-full lg:w-1/2">
-        <div class="flex justify-center lg:justify-start">
-          <p class=" my-2 max-w-xl py-6">{{about}}</p>
-        </div>
-      </div>
-  </div>
-</div>
