+++ src/app/page-sections/hero/hero.component.css
@@ -1,0 +1,16 @@
+.heroHighLights {
+  margin-right: 20px;
+  /* Gradient border */
+  border: 2px solid transparent;
+  background-image: linear-gradient(#333, #333),
+                    linear-gradient(to right, #ec4899, #64748b, #a855f7);
+  background-origin: border-box;
+  background-clip: padding-box, border-box;
+
+  transition: background-color 0.3s ease, transform 0.3s ease;
+}
+
+.heroHighLights:hover{
+  transform: scale(1.05);
+  color: white;
+}
