+++ src/app/page-sections/hero/hero.component.html
@@ -2,8 +2,14 @@
   <div class="flex flex-wrap">
     <div class="w-full lg:w-1/2">
      <div class="flex flex-col items-center lg:items-start">
-      <h1 class="pb-16 text-4xl font-thin tracking-tight lg:mt-10 lg:text-6xl">Benyakoub M.</h1>
-      <span class="bg-gradient-to-r from-pink-300 via-slate-500 to-purple-500 bg-clip-text text-2xl tracking-tight text-transparent"> Senior Web FrontEnd Consultant (Angular)</span>
+      <h1 class="pb-4 text-4xl font-thin tracking-tight lg:mt-10 lg:text-6xl">Benyakoub M.</h1>
+      <span class="pl-1 text-center bg-gradient-to-r from-pink-300 via-slate-500 to-purple-500 bg-clip-text text-2xl tracking-tight text-transparent"> Senior Web FrontEnd Consultant (Angular)</span>
+            <!-- Download CV Button -->
+      <div class="pl-1 flex justify-center lg:justify-start">
+        <a href="/assets/files/benResume.pdf" download="BenResume.pdf" class=" mt-3 heroHighLights rounded-xl border-4 border-neutral-800 p-4 bg-neutral-800 text-white hover:bg-white hover:text-black transition-all duration-300 ease-in-out">
+          Download My CV
+        </a>
+      </div>
       <p class="my-2 max-w-xl py-6 font-light tracking-tight">{{ heroContent }}</p>
      </div>
     </div>
@@ -26,7 +32,7 @@
     </div>
 
     <!-- Box 3 -->
-    <div class="heroHighLights heroHighLights rounded-2xl border-4 border-neutral-800 p-6 bg-neutral-800 text-white hover:bg-white hover:text-black transition-all duration-300 ease-in-out">
+    <div class="heroHighLights rounded-2xl border-4 border-neutral-800 p-6 bg-neutral-800 text-white hover:bg-white hover:text-black transition-all duration-300 ease-in-out">
       <h4>8+ Technologies Mastered</h4>
     </div>
