+++ src/app/page-sections/projects/projects.component.html
@@ -16,5 +16,5 @@
     </div>
   </div>
 
-  <h3 class="w-40 h-50 my-20 text-center text-2xl text-white font-semibold rounded-md hover:bg-neutral-800 transition duration-300 ease-in-out"><a href="https://www.linkedin.com/in/benyakoub/" target="_blank">See More <i class="bi bi-arrow-right cursor-pointer"> </i></a> </h3>
+  <h3 class=" w-40 h-50 mb-20 text-center text-2xl bg-neutral-900 text-white font-semibold rounded-md hover:bg-neutral-950 transition duration-300 ease-in-out"><a href="https://www.linkedin.com/in/benyakoub/" target="_blank">See More <i class="bi bi-arrow-right cursor-pointer"> </i></a> </h3>
 </div>
