+++ src/app/page-sections/projects/projects.component.ts
