+++ src/app/page-sections/technologies/technologies.component.ts
